==> notebooks/3-DATA_TREATMENT_CREATING_FEATURES.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import geopandas as gpd
import shapely.geometry
from shapely.geometry import Point
from mpl_toolkits.basemap import Basemap
import glob
import rasterio
import os
import shutil
import pyimpute
from pathlib import Path

BIO_PATH = "assets/wc2.1_30s_bio/*.tif"
FILE_PATH = "generated_files\gbif.parquet"
raster_features = sorted(glob.glob(BIO_PATH))

# Ocurrence species data -> OSD_df
OSD_df = pd.read_parquet(FILE_PATH)
OSD_df['geometry'] = list(zip(OSD_df["Longitude"], OSD_df["Latitude"]))
OSD_df['geometry'] = OSD_df["geometry"].apply(Point)

# Create the geodataframe
OSD_geoframe = gpd.GeoDataFrame(
    OSD_df,
    crs = {'init': 'epsg:4326'},
    geometry = OSD_df['geometry']
)
OSD_geoframe = OSD_geoframe.to_crs("EPSG:4326")
OSD_geoframe.reset_index(drop=True, inplace = True)

coord_list = [(x,y) for x,y in zip(OSD_geoframe['geometry'].x , OSD_geoframe['geometry'].y)]

# Point sampling
for f in raster_features:
    src = rasterio.open(f)
    OSD_df[Path(f).stem] = [x for x in src.sample(coord_list)]
    OSD_df[Path(f).stem] = OSD_df[Path(f).stem].astype('float64')

COLUMNS_RENAME_DICT = {
    'wc2.1_30s_bio_1':'01_annual_mean_temp',
    'wc2.1_30s_bio_2':'02_mean_diurnal_range',
    'wc2.1_30s_bio_3':'03_isothermality',
    'wc2.1_30s_bio_4':'04_temperature_seasonality',
    'wc2.1_30s_bio_5':'05_maximum_temp_warmest_month',
    'wc2.1_30s_bio_6':'06_minimum_temp_coldest_month',
    'wc2.1_30s_bio_7':'07_temp_annual_range',
    'wc2.1_30s_bio_8':'08_mean_temp_wettest_quarter',
    'wc2.1_30s_bio_9':'09_mean_temp_driest_quarter',
    'wc2.1_30s_bio_10':'10_mean_temp_warmest_quarter',
    'wc2.1_30s_bio_11':'11_mean_temp_coldest_quarter',
    'wc2.1_30s_bio_12':'12_annual_precipitation',
    'wc2.1_30s_bio_13':'13_precipitation_wettest_month',
    'wc2.1_30s_bio_14':'14_precipitation_driest_month',
    'wc2.1_30s_bio_15':'15_precipitation_seasonality',
    'wc2.1_30s_bio_16':'16_precipitation_of_wettest_quarter',
    'wc2.1_30s_bio_17':'17_precipitation_of_driest_quarter',
    'wc2.1_30s_bio_18':'18_precipitation_of_warmest_quarter',
    'wc2.1_30s_bio_19':'19_precipitation_of_coldest_quarter',
}

OSD_df.rename(columns=COLUMNS_RENAME_DICT, inplace=True)

OSD_df

OSD_df.drop(columns=["geometry"]).to_parquet("generated_files/bio_variables_dataframe_occurence.parquet", index=False)
OSD_df.drop(columns=["geometry"]).to_csv("generated_files/bio_variables_dataframe_occurence.csv", index=False)

from urllib import request
import json

STATES = ["RS","SC","PR","SP","MG","ES","RJ","MS"]
url = 'https://raw.githubusercontent.com/codeforgermany/click_that_hood/main/public/data/brazil-states.geojson'
with request.urlopen(url) as f:
    brazil_states = json.load(f)
    
brazil_states_filtered = [feat for feat in brazil_states['features'] if feat['properties']['sigla'] in STATES]
updated_brazil_states_filtered = {'type':brazil_states['type'],'features':brazil_states_filtered}

coordinates_list = []

for feats in updated_brazil_states_filtered['features']:
    coordinates_list += feats['geometry']['coordinates'][0][0]

dataframe = pd.DataFrame(coordinates_list, columns=["Longitude","Latitude"])

# Point sampling
for f in raster_features:
    src = rasterio.open(f)
    dataframe[Path(f).stem] = [x for x in src.sample(coordinates_list)]
    dataframe[Path(f).stem] = dataframe[Path(f).stem].astype('float64')

dataframe.rename(columns=COLUMNS_RENAME_DICT, inplace=True)

dataframe.to_parquet("generated_files/bio_variables_dataframe_base.parquet", index=False)
dataframe.to_csv("generated_files/bio_variables_dataframe_base.csv", index=False)

# Utilizando somente a feature 1
brazil = gpd.read_file("assets\BR_UF_2022\BR_UF_2022.shp")

brazil = brazil[brazil['SIGLA_UF'].isin(STATES)]

polygon = brazil['geometry'].unary_union

from rasterio.features import shapes
from shapely.geometry import shape, Polygon, Point

with rasterio.open('assets\wc2.1_30s_bio\wc2.1_30s_bio_1.tif') as src:
    # Extrair os metadados
    profile = src.profile
    # Ler os dados raster
    array = src.read(1)

# Converter o polígono para um objeto Shapely
polygon_shapely = shape(polygon)

# Identificar as formas e transformá-las em polígonos
polygons = [polygon_shapely]

mask = rasterio.features.geometry_mask(polygons, out_shape=array.shape, transform=profile['transform'], invert=True)

# Coletar os pontos dentro do polígono
points_within_polygon = []
for idx, val in np.ndenumerate(mask):
    if not val:
        row, col = idx
        lon, lat = src.xy(row, col)
        points_within_polygon.append((lon, lat))

print("Pontos dentro do polígono:", points_within_polygon[:3])

dataframe = pd.DataFrame(points_within_polygon, columns=["Longitude","Latitude"])

dataframe.to_parquet("generated_files/bio_variables_dataframe_base_var_1.parquet", index=False)
dataframe.to_csv("generated_files/bio_variables_dataframe_base_var_1.csv", index=False)
