==> notebooks/10-PLOTTING_CORRELATION.ipynb <==
import geopandas as gpd
import pandas as pd
from scripts.load_gbif import get_dataframe
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

dataframe = get_dataframe()

dataframe = pd.read_parquet("generated_files/bio_variables_dataframe_base.parquet")

min_mask = dataframe['01_annual_mean_temp'].min()
dataframe = dataframe[dataframe['01_annual_mean_temp'] > min_mask][dataframe.columns[2:]].copy()

renamed_cols = [i[:2] for i in dataframe.columns]
dataframe.columns = renamed_cols
dataframe = dataframe[sorted(dataframe.columns)]

df_corr= dataframe.corr()

import plotly.express as px


fig = px.imshow(np.abs(df_corr))
fig.show()

columns = ['01','02','07','12','14','18']

df_corr_2= dataframe[columns].corr()

fig = px.imshow(np.abs(df_corr_2))
fig.show()

temperature_cols = [
    '01_annual_mean_temp',
    '02_mean_diurnal_range',
    '03_isothermality',
    '04_temperature_seasonality'
    ]
temperature_cols_2 = [
    '05_maximum_temp_warmest_month',
    '06_minimum_temp_coldest_month',
    '07_temp_annual_range',
    '08_mean_temp_wettest_quarter',
    '09_mean_temp_driest_quarter',
    '10_mean_temp_warmest_quarter',
    '11_mean_temp_coldest_quarter'
    ]
precipitation_cols = [
    '12_annual_precipitation',
    '13_precipitation_wettest_month',
    '14_precipitation_driest_month',
    '15_precipitation_seasonality',
    '16_precipitation_of_wettest_quarter',
    '17_precipitation_of_driest_quarter',
    '18_precipitation_of_warmest_quarter',
    '19_precipitation_of_coldest_quarter'
    ]

dataframe = pd.read_parquet("generated_files/bio_variables_dataframe_base.parquet")

selected_cols = ['01_annual_mean_temp','02_mean_diurnal_range','07_temp_annual_range','12_annual_precipitation','14_precipitation_driest_month','18_precipitation_of_warmest_quarter']
selected_cols2 = ['01_annual_mean_temp','02_mean_diurnal_range','07_temp_annual_range','12_annual_precipitation']

sns.pairplot(dataframe[dataframe['01_annual_mean_temp'] > min_mask][selected_cols2])
plt.show()
plt.savefig("generated_files/selected_cols2.png")

min_mask = dataframe[temperature_cols]['01_annual_mean_temp'].min()

sns.pairplot(dataframe[dataframe['01_annual_mean_temp'] > min_mask][temperature_cols_2])
plt.show()
plt.savefig("generated_files/temp.png")

sns.pairplot(dataframe[dataframe['01_annual_mean_temp'] > min_mask][temperature_cols])
plt.show()
plt.savefig("generated_files/mean_temp.png")

sns.pairplot(dataframe[dataframe['01_annual_mean_temp'] > min_mask][precipitation_cols])
plt.show()
plt.savefig("generated_files/precipitation.png")
